Move test files to a shared project for reuse
Move TinyIoC and external types to shared projects so they can be easily referenced from .NET Core apps and .NET Standardard class libraries, with the exception of TinyIoCAspNetExtensions, which is not compatible with .NET Core in its current form due to its use of HttpContext.Current. Create a .NET Standard libraries to build both TinyIoC and the external types.